+++ nntrainer/app_context.cpp
@@ -238,171 +238,171 @@ static void add_default_object(AppContext &ac) {
   /// @note all layers should be added to the app_context to guarantee that
   /// createLayer/createOptimizer class is created
   using OptType = ml::train::OptimizerType;
-  ac.registerFactory(nntrainer::createOptimizer<SGD>, SGD::type, OptType::SGD);
+  ac.registerFactory(nntrainer::createOptimizer<SGD>, SGD::type, static_cast<int>(OptType::SGD));
   ac.registerFactory(nntrainer::createOptimizer<Adam>, Adam::type,
-                     OptType::ADAM);
+    static_cast<int>(OptType::ADAM));
   ac.registerFactory(nntrainer::createOptimizer<AdamW>, AdamW::type,
-                     OptType::ADAMW);
+    static_cast<int>(OptType::ADAMW));
   ac.registerFactory(AppContext::unknownFactory<nntrainer::Optimizer>,
-                     "unknown", OptType::UNKNOWN);
+                     "unknown", static_cast<int>(OptType::UNKNOWN));
 
   using LRType = LearningRateSchedulerType;
   ac.registerFactory(
     ml::train::createLearningRateScheduler<ConstantLearningRateScheduler>,
-    ConstantLearningRateScheduler::type, LRType::CONSTANT);
+    ConstantLearningRateScheduler::type, static_cast<int>(LRType::CONSTANT));
   ac.registerFactory(
     ml::train::createLearningRateScheduler<ExponentialLearningRateScheduler>,
-    ExponentialLearningRateScheduler::type, LRType::EXPONENTIAL);
+    ExponentialLearningRateScheduler::type, static_cast<int>(LRType::EXPONENTIAL));
   ac.registerFactory(
     ml::train::createLearningRateScheduler<StepLearningRateScheduler>,
-    StepLearningRateScheduler::type, LRType::STEP);
+    StepLearningRateScheduler::type, static_cast<int>(LRType::STEP));
   ac.registerFactory(ml::train::createLearningRateScheduler<
                        CosineAnnealingLearningRateScheduler>,
                      CosineAnnealingLearningRateScheduler::type,
-                     LRType::COSINE);
+                     static_cast<int>(LRType::COSINE));
   ac.registerFactory(
     ml::train::createLearningRateScheduler<LinearLearningRateScheduler>,
-    LinearLearningRateScheduler::type, LRType::LINEAR);
+    LinearLearningRateScheduler::type, static_cast<int>(LRType::LINEAR));
 
   using LayerType = ml::train::LayerType;
   ac.registerFactory(nntrainer::createLayer<InputLayer>, InputLayer::type,
-                     LayerType::LAYER_IN);
+                     static_cast<int>(LayerType::LAYER_IN));
   ac.registerFactory(nntrainer::createLayer<WeightLayer>, WeightLayer::type,
-                     LayerType::LAYER_WEIGHT);
+                     static_cast<int>(LayerType::LAYER_WEIGHT));
   ac.registerFactory(nntrainer::createLayer<AddLayer>, AddLayer::type,
-                     LayerType::LAYER_ADD);
+                     static_cast<int>(LayerType::LAYER_ADD));
   ac.registerFactory(nntrainer::createLayer<SubtractLayer>, SubtractLayer::type,
-                     LayerType::LAYER_SUBTRACT);
+                     static_cast<int>(LayerType::LAYER_SUBTRACT));
   ac.registerFactory(nntrainer::createLayer<MultiplyLayer>, MultiplyLayer::type,
-                     LayerType::LAYER_MULTIPLY);
+                     static_cast<int>(LayerType::LAYER_MULTIPLY));
   ac.registerFactory(nntrainer::createLayer<DivideLayer>, DivideLayer::type,
-                     LayerType::LAYER_DIVIDE);
+                     static_cast<int>(LayerType::LAYER_DIVIDE));
   ac.registerFactory(nntrainer::createLayer<PowLayer>, PowLayer::type,
-                     LayerType::LAYER_POW);
+                     static_cast<int>(LayerType::LAYER_POW));
   ac.registerFactory(nntrainer::createLayer<SQRTLayer>, SQRTLayer::type,
-                     LayerType::LAYER_SQRT);
+                     static_cast<int>(LayerType::LAYER_SQRT));
   ac.registerFactory(nntrainer::createLayer<SineLayer>, SineLayer::type,
-                     LayerType::LAYER_SINE);
+                     static_cast<int>(LayerType::LAYER_SINE));
   ac.registerFactory(nntrainer::createLayer<CosineLayer>, CosineLayer::type,
-                     LayerType::LAYER_COSINE);
+                     static_cast<int>(LayerType::LAYER_COSINE));
   ac.registerFactory(nntrainer::createLayer<TangentLayer>, TangentLayer::type,
-                     LayerType::LAYER_TANGENT);
+                     static_cast<int>(LayerType::LAYER_TANGENT));
   ac.registerFactory(nntrainer::createLayer<MatMulLayer>, MatMulLayer::type,
-                     LayerType::LAYER_MATMUL);
+                     static_cast<int>(LayerType::LAYER_MATMUL));
   ac.registerFactory(nntrainer::createLayer<NegLayer>, NegLayer::type,
-                     LayerType::LAYER_NEG);
+                     static_cast<int>(LayerType::LAYER_NEG));
   ac.registerFactory(nntrainer::createLayer<FullyConnectedLayer>,
-                     FullyConnectedLayer::type, LayerType::LAYER_FC);
+                     FullyConnectedLayer::type, static_cast<int>(LayerType::LAYER_FC));
   ac.registerFactory(nntrainer::createLayer<BatchNormalizationLayer>,
-                     BatchNormalizationLayer::type, LayerType::LAYER_BN);
+                     BatchNormalizationLayer::type, static_cast<int>(LayerType::LAYER_BN));
   ac.registerFactory(nntrainer::createLayer<LayerNormalizationLayer>,
                      LayerNormalizationLayer::type,
-                     LayerType::LAYER_LAYER_NORMALIZATION);
+                     static_cast<int>(LayerType::LAYER_LAYER_NORMALIZATION));
   ac.registerFactory(nntrainer::createLayer<Conv2DLayer>, Conv2DLayer::type,
-                     LayerType::LAYER_CONV2D);
+                     static_cast<int>(LayerType::LAYER_CONV2D));
   ac.registerFactory(nntrainer::createLayer<Conv2DTransposeLayer>,
                      Conv2DTransposeLayer::type,
-                     LayerType::LAYER_CONV2D_TRANSPOSE);
+                     static_cast<int>(LayerType::LAYER_CONV2D_TRANSPOSE));
   ac.registerFactory(nntrainer::createLayer<Conv1DLayer>, Conv1DLayer::type,
-                     LayerType::LAYER_CONV1D);
+                     static_cast<int>(LayerType::LAYER_CONV1D));
   ac.registerFactory(nntrainer::createLayer<Pooling2DLayer>,
-                     Pooling2DLayer::type, LayerType::LAYER_POOLING2D);
+                     Pooling2DLayer::type, static_cast<int>(LayerType::LAYER_POOLING2D));
   ac.registerFactory(nntrainer::createLayer<FlattenLayer>, FlattenLayer::type,
-                     LayerType::LAYER_FLATTEN);
+                     static_cast<int>(LayerType::LAYER_FLATTEN));
   ac.registerFactory(nntrainer::createLayer<ReshapeLayer>, ReshapeLayer::type,
-                     LayerType::LAYER_RESHAPE);
+                     static_cast<int>(LayerType::LAYER_RESHAPE));
   ac.registerFactory(nntrainer::createLayer<ActivationLayer>,
-                     ActivationLayer::type, LayerType::LAYER_ACTIVATION);
+                     ActivationLayer::type, static_cast<int>(LayerType::LAYER_ACTIVATION));
   ac.registerFactory(nntrainer::createLayer<AdditionLayer>, AdditionLayer::type,
-                     LayerType::LAYER_ADDITION);
+                     static_cast<int>(LayerType::LAYER_ADDITION));
   ac.registerFactory(nntrainer::createLayer<ConcatLayer>, ConcatLayer::type,
-                     LayerType::LAYER_CONCAT);
+                     static_cast<int>(LayerType::LAYER_CONCAT));
   ac.registerFactory(nntrainer::createLayer<MultiOutLayer>, MultiOutLayer::type,
-                     LayerType::LAYER_MULTIOUT);
+                     static_cast<int>(LayerType::LAYER_MULTIOUT));
   ac.registerFactory(nntrainer::createLayer<EmbeddingLayer>,
-                     EmbeddingLayer::type, LayerType::LAYER_EMBEDDING);
+                     EmbeddingLayer::type, static_cast<int>(LayerType::LAYER_EMBEDDING));
   ac.registerFactory(nntrainer::createLayer<RNNLayer>, RNNLayer::type,
-                     LayerType::LAYER_RNN);
+                     static_cast<int>(LayerType::LAYER_RNN));
   ac.registerFactory(nntrainer::createLayer<RNNCellLayer>, RNNCellLayer::type,
-                     LayerType::LAYER_RNNCELL);
+                     static_cast<int>(LayerType::LAYER_RNNCELL));
   ac.registerFactory(nntrainer::createLayer<LSTMLayer>, LSTMLayer::type,
-                     LayerType::LAYER_LSTM);
+                     static_cast<int>(LayerType::LAYER_LSTM));
   ac.registerFactory(nntrainer::createLayer<LSTMCellLayer>, LSTMCellLayer::type,
-                     LayerType::LAYER_LSTMCELL);
+                     static_cast<int>(LayerType::LAYER_LSTMCELL));
   ac.registerFactory(nntrainer::createLayer<ZoneoutLSTMCellLayer>,
                      ZoneoutLSTMCellLayer::type,
-                     LayerType::LAYER_ZONEOUT_LSTMCELL);
+                     static_cast<int>(LayerType::LAYER_ZONEOUT_LSTMCELL));
   ac.registerFactory(nntrainer::createLayer<SplitLayer>, SplitLayer::type,
-                     LayerType::LAYER_SPLIT);
+                     static_cast<int>(LayerType::LAYER_SPLIT));
   ac.registerFactory(nntrainer::createLayer<GRULayer>, GRULayer::type,
-                     LayerType::LAYER_GRU);
+                     static_cast<int>(LayerType::LAYER_GRU));
   ac.registerFactory(nntrainer::createLayer<GRUCellLayer>, GRUCellLayer::type,
-                     LayerType::LAYER_GRUCELL);
+                     static_cast<int>(LayerType::LAYER_GRUCELL));
   ac.registerFactory(nntrainer::createLayer<PermuteLayer>, PermuteLayer::type,
-                     LayerType::LAYER_PERMUTE);
+                     static_cast<int>(LayerType::LAYER_PERMUTE));
   ac.registerFactory(nntrainer::createLayer<DropOutLayer>, DropOutLayer::type,
-                     LayerType::LAYER_DROPOUT);
+                     static_cast<int>(LayerType::LAYER_DROPOUT));
   ac.registerFactory(nntrainer::createLayer<AttentionLayer>,
-                     AttentionLayer::type, LayerType::LAYER_ATTENTION);
+                     AttentionLayer::type, static_cast<int>(LayerType::LAYER_ATTENTION));
   ac.registerFactory(nntrainer::createLayer<MoLAttentionLayer>,
-                     MoLAttentionLayer::type, LayerType::LAYER_MOL_ATTENTION);
+                     MoLAttentionLayer::type, static_cast<int>(LayerType::LAYER_MOL_ATTENTION));
   ac.registerFactory(nntrainer::createLayer<MultiHeadAttentionLayer>,
                      MultiHeadAttentionLayer::type,
-                     LayerType::LAYER_MULTI_HEAD_ATTENTION);
+                     static_cast<int>(LayerType::LAYER_MULTI_HEAD_ATTENTION));
   ac.registerFactory(nntrainer::createLayer<ReduceMeanLayer>,
-                     ReduceMeanLayer::type, LayerType::LAYER_REDUCE_MEAN);
+                     ReduceMeanLayer::type, static_cast<int>(LayerType::LAYER_REDUCE_MEAN));
   ac.registerFactory(nntrainer::createLayer<PositionalEncodingLayer>,
                      PositionalEncodingLayer::type,
-                     LayerType::LAYER_POSITIONAL_ENCODING);
+                     static_cast<int>(LayerType::LAYER_POSITIONAL_ENCODING));
   ac.registerFactory(nntrainer::createLayer<IdentityLayer>, IdentityLayer::type,
-                     LayerType::LAYER_IDENTITY);
+                     static_cast<int>(LayerType::LAYER_IDENTITY));
   ac.registerFactory(nntrainer::createLayer<Upsample2dLayer>,
-                     Upsample2dLayer::type, LayerType::LAYER_UPSAMPLE2D);
+                     Upsample2dLayer::type, static_cast<int>(LayerType::LAYER_UPSAMPLE2D));
 
   ac.registerFactory(nntrainer::createLayer<ChannelShuffle>,
-                     ChannelShuffle::type, LayerType::LAYER_CHANNEL_SHUFFLE);
+                     ChannelShuffle::type, static_cast<int>(LayerType::LAYER_CHANNEL_SHUFFLE));
 
 #ifdef ENABLE_NNSTREAMER_BACKBONE
   ac.registerFactory(nntrainer::createLayer<NNStreamerLayer>,
                      NNStreamerLayer::type,
-                     LayerType::LAYER_BACKBONE_NNSTREAMER);
+                     static_cast<int>(LayerType::LAYER_BACKBONE_NNSTREAMER));
 #endif
 #ifdef ENABLE_TFLITE_BACKBONE
   ac.registerFactory(nntrainer::createLayer<TfLiteLayer>, TfLiteLayer::type,
-                     LayerType::LAYER_BACKBONE_TFLITE);
+                     static_cast<int>(LayerType::LAYER_BACKBONE_TFLITE));
 #endif
   ac.registerFactory(nntrainer::createLayer<CentroidKNN>, CentroidKNN::type,
-                     LayerType::LAYER_CENTROID_KNN);
+                     static_cast<int>(LayerType::LAYER_CENTROID_KNN));
 
   /** preprocess layers */
   ac.registerFactory(nntrainer::createLayer<PreprocessFlipLayer>,
                      PreprocessFlipLayer::type,
-                     LayerType::LAYER_PREPROCESS_FLIP);
+                     static_cast<int>(LayerType::LAYER_PREPROCESS_FLIP));
   ac.registerFactory(nntrainer::createLayer<PreprocessTranslateLayer>,
                      PreprocessTranslateLayer::type,
-                     LayerType::LAYER_PREPROCESS_TRANSLATE);
+                     static_cast<int>(LayerType::LAYER_PREPROCESS_TRANSLATE));
   ac.registerFactory(nntrainer::createLayer<PreprocessL2NormLayer>,
                      PreprocessL2NormLayer::type,
-                     LayerType::LAYER_PREPROCESS_L2NORM);
+                     static_cast<int>(LayerType::LAYER_PREPROCESS_L2NORM));
 
   /** register losses */
   ac.registerFactory(nntrainer::createLayer<MSELossLayer>, MSELossLayer::type,
-                     LayerType::LAYER_LOSS_MSE);
+                     static_cast<int>(LayerType::LAYER_LOSS_MSE));
   ac.registerFactory(nntrainer::createLayer<CrossEntropySigmoidLossLayer>,
                      CrossEntropySigmoidLossLayer::type,
-                     LayerType::LAYER_LOSS_CROSS_ENTROPY_SIGMOID);
+                     static_cast<int>(LayerType::LAYER_LOSS_CROSS_ENTROPY_SIGMOID));
   ac.registerFactory(nntrainer::createLayer<CrossEntropySoftmaxLossLayer>,
                      CrossEntropySoftmaxLossLayer::type,
-                     LayerType::LAYER_LOSS_CROSS_ENTROPY_SOFTMAX);
+                     static_cast<int>(LayerType::LAYER_LOSS_CROSS_ENTROPY_SOFTMAX));
   ac.registerFactory(nntrainer::createLayer<ConstantDerivativeLossLayer>,
                      ConstantDerivativeLossLayer::type,
-                     LayerType::LAYER_LOSS_CONSTANT_DERIVATIVE);
+                     static_cast<int>(LayerType::LAYER_LOSS_CONSTANT_DERIVATIVE));
 
   ac.registerFactory(nntrainer::createLayer<TimeDistLayer>, TimeDistLayer::type,
-                     LayerType::LAYER_TIME_DIST);
+                     static_cast<int>(LayerType::LAYER_TIME_DIST));
 
   ac.registerFactory(AppContext::unknownFactory<nntrainer::Layer>, "unknown",
-                     LayerType::LAYER_UNKNOWN);
+                     static_cast<int>(LayerType::LAYER_UNKNOWN));
 }
 
 static void add_extension_object(AppContext &ac) {
